config/db.py
import os
from dotenv import load_dotenv
import mysql.connector

def connectdb():
    host = os.getenv('DB_HOST')
    user = os.getenv('DB_USER')
    password = os.getenv('DB_PASS')
    database = os.getenv('DB_NAME')
    port = os.getenv('DB_PORT')
    return mysql.connector.connect(
        host=host,
        user=user,
        password=password,
        database=database,
        port=port
    )